# Use PkScan operator