--- a/db_test.go
+++ b/db_test.go
@@ -163,3 +163,26 @@ func BenchmarkSelectWhere(b *testing.B) {
 		})
 	}
 }
+
+func BenchmarkSelectPk(b *testing.B) {
+	for size := 1; size <= 10000; size *= 10 {
+		b.Run(fmt.Sprintf("%.05d", size), func(b *testing.B) {
+			db, err := genji.Open(":memory:")
+			require.NoError(b, err)
+
+			err = db.Exec("CREATE TABLE foo(a INT PRIMARY KEY)")
+			require.NoError(b, err)
+
+			for i := 0; i < size; i++ {
+				err = db.Exec("INSERT INTO foo(a) VALUES (?)", i)
+				require.NoError(b, err)
+			}
+
+			b.ResetTimer()
+			for i := 0; i < b.N; i++ {
+				res, _ := db.Query("SELECT * FROM foo WHERE a = ?", size-1)
+				res.Iterate(func(d document.Document) error { return nil })
+			}
+		})
+	}
+}